Add gulp file to build stylesheet + test demo
Since we've moved the source icons to an image set (fticons) we need a
gulp task that will generate a stylesheet of icons so that these icons
are available with the build service.

This gulp task expects a file called imageList.json to exist in fticons
which it then uses to create a Sass list in _icon-list.scss.

This gulp task also stores a copy of the imageList to data.json so that
a local test of the icons can be run