Delete icons demo
The interesting bit of o-icons (the icons) have moved to a seperate
repo for images. This means we can remove the visual demo for the
registry.

However it is useful to have a test demo to run locally to check
the icons are working so I've created a test.mustache demo for that
purpose.